+++ src/cosmos/cosmosUtils.ts
@@ -1,4 +1,4 @@
-import { getIbcInfo, getTransferChannel } from '@chain-registry/utils'
+import { getIBCData, getTransferChannel } from '@chain-registry/utils'
 import { addCoins, StdSignDoc } from '@cosmjs/amino'
 import { fromBech32 } from '@cosmjs/encoding'
 import {
@@ -26,7 +26,7 @@ import {
   Tendermint37Client
 } from '@cosmjs/tendermint-rpc'
 import { add } from 'biggystring'
-import { chains, ibc } from 'chain-registry'
+import { chains, ibcData } from 'chain-registry'
 import { asMaybe, asObject, asString, asTuple, asValue } from 'cleaners'
 import { EdgeFetchFunction } from 'edge-core-js/types'
 import { base16, base64 } from 'rfc4648'
@@ -265,12 +265,12 @@ export const getIbcChannelAndPort = (
   toAddress: string
 ): IbcChannel => {
   const chain = chains.find(
-    chain => chain.bech32_prefix === fromBech32(toAddress).prefix
+    chain => chain.bech32Prefix === fromBech32(toAddress).prefix
   )
   if (chain == null) throw new Error('Unrecognized denom')
 
-  const toChainName = chain.chain_name
-  const ibcInfo = getIbcInfo(ibc, fromChainName, toChainName)
+  const toChainName = chain.chainName
+  const ibcInfo = getIBCData(ibcData, fromChainName, toChainName)
   if (ibcInfo == null) {
     throw new Error(
       `No IBC channels between ${fromChainName} and ${toChainName}`
@@ -282,13 +282,13 @@ export const getIbcChannelAndPort = (
   // channel data is alphabetical by chain name
   if (fromChainName < toChainName) {
     return {
-      channel: channel.chain_1.channel_id,
-      port: channel.chain_1.port_id
+      channel: channel.chain1.channelId,
+      port: channel.chain1.portId
     }
   } else {
     return {
-      channel: channel.chain_2.channel_id,
-      port: channel.chain_2.port_id
+      channel: channel.chain2.channelId,
+      port: channel.chain2.portId
     }
   }
 }